fileStream: Init the size of files.
Init the size of new files, otherwise it will process the whole file
at the first interval.

We have also add a threshold for old files. A file will be classified as
old file when it's not updated in the lastest interval, we think it is
irreasonable, so we add a oldThreshold argument.